[IMP] account: post total levies
In certain countries or industries, practice goes to retain certain invoice
amounts ("levies" or "guarantees") after the invoice total.
Those instruments do not financially alter the value of the transaction, but
corresponding funds are not directly settled between the parties, but rather
through a government account or some escrow or similar.

This scenario is represented by this PR. As this is only relevent on invoicing,
but absolute irrelevant for sales or purchasing activiy, those ignore
post total levies. Only on the invoice document, those are exposed.