[FIX] purchase: align invoice stats with SO
_set_additional_fields is called on self (invoice) and therefore (can) makes
use of self's state.

However, during the purchase onchange, the state of self can diverge from
a comparative state of sale order.
This is due to tha fact, that state in sales order is create-tie state,
while state in purchase is an undeterministic onchange state.

In the case, that partner onchange influences relevant state
for _set_additional_fields (eg. additional fields are set based on fiscal
position, which in turn is set by partner onchange), this commit now provides a
consistent state to _set_additional_fields analogue to the state in sale module.