[IMP] crm: do allow merging with 100% ops
When I'm done with an opportunity, it might regularily happen
that the client does give me some late feedback on this opportunity
via mail.

When cleaning up my sales pipeline I want to be able to organize this
feedback into the right history: the original opportunity.

This commit allows me to merge opportunities with won opportunities, too.
As this is optional, the previous semantic can be preserved without further
effort.