[IMP] point_of_sale: make2order on a pos pickings
Before this commit, pos_restaurant could not be configured to trigger a bom
explosion of a freshly prepared dish, in order to track ongoing (estimated)
raw material consumtion (estimated, as based on estimated bom values).

This commit replays the logic employed in mrp module onto the pos module,
that is: when the route that happens to reflect the specs of a move in the
picking at hand defines make_to_order, adjust that move accordingly.

After this, a product or product category can be configured (through a route) to
be make_to_order.

Apart from the above use case, this also enables routing logic for a
Point of Sale, if - although a POS - the business requires it. Example:
Dispense of electronic goods often times have a separate dispatch window, need
a serial registration an IMEI unlock, etc. while still beeing sold through a
POS sales flow.