[IMP] account: add CoA update wizard
This allows to manage CoA extensions through third party modules,
thos modules would run a post init hook calling the _install_template method
with an approriate context.

It also provides a wizard for the end user to load missing and or update a
Chart of Account. For speed, the loading can be limited to the scope of a
single module. This single module mode would also be the preferred option for
post init hooks.

As an advanced feature, the wizard is only exposed as technical setting.

Properties and journals are excemt from this loading mode, as they do not
have tamplate data. Therefore, they are impossible to track. There does not seem
to be a valid use case to update those, neither.

Signed-off-by: "David Arnold" <"dar@xoe.solutions">