[FIX] pos: Receipt T-seperator and line break issues
The thousand separator is not set by default, but a method is
available to render it according to the currency setting.

Use that method.

Furthermore the tax details easily drop out of the <left> space,
which is liited to 24 characters. A complex line wrapper could
be created. Instead just add a line break which is save in
english wording up to a base amount of 9.999.999,99 - enough.