[IMP] account: prevent mistakes using replace statement for sql queries
In the reconciliation process, the name of the tables ["res_partner","res_partner_bank"]
is replaced by an alias through the python method `replace`, but it replaces all found matches
for the tables name in a sql sentence. This implementation is sensible to make
wrong replacements (sub-queries and possible tables/columns that contains the tables name).

Therefore, replacement needs be more strict according to sql syntax 
adding dot at the end of each table's name like an identifier
(signal when a table is accessing to one of its columns), this is a safer point 
where the table's name can be replaced.